third_party/WebKit/LayoutTests/fast/imagecapture/ImageCapture-creation.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script src=../../resources/testharness-helpers.js></script>
<script>

// This test verifies that ImageCapture can be created (or not) with different
// Media Stream Track types (audio, video).

var testVideo = promise_test(function() {
  const gotStream = this.step_func(function(stream) {
    assert_equals(stream.getAudioTracks().length, 0);
    assert_equals(stream.getVideoTracks().length, 1);
    assert_equals(stream.getVideoTracks()[0].readyState, 'live');
    assert_true(stream.getVideoTracks()[0].enabled);
    assert_false(stream.getVideoTracks()[0].muted);

    var capturer = new ImageCapture(stream.getVideoTracks()[0]);
    assert_equals(capturer.videoStreamTrack, stream.getVideoTracks()[0]);

    // Assert that grabFrame() is rejected if the associated video track is
    // disabled, or ended. grabFrame() would also reject if the video Track is
    // muted but that's a read-only property.
    stream.getVideoTracks()[0].enabled = false;
    assert_promise_rejects(capturer.grabFrame(),
                           'InvalidStateError',
                           'ImageCapturer cannot grabFrame() of a disabled Track');

    stream.getVideoTracks()[0].stop();
    assert_equals(stream.getVideoTracks()[0].readyState, 'ended');
    assert_promise_rejects(capturer.grabFrame(),
                           'InvalidStateError',
                           'ImageCapturer cannot grabFrame() of a non-live Track');

    this.done();
  });

  const onError = this.step_func(function() {
    assert_unreached('Error creating MediaStream');
  });

  navigator.webkitGetUserMedia({video:true}, gotStream, onError);
}, 'verifies that ImageCapture API rejects grabFrame() in certain Track states.');

var testAudio = promise_test(function() {
  const onError = this.step_func(function() {
     assert_unreached('Error creating MediaStream');
  });

  navigator.webkitGetUserMedia({audio:true},
    this.step_func(function(stream) {
      assert_equals(stream.getAudioTracks().length, 1);
      assert_equals(stream.getVideoTracks().length, 0);
      assert_throws("NotSupportedError",
                    function() {
                      var capturer = new ImageCapture(stream.getAudioTracks()[0]);
                    },
                    'an ImageCapturer can only be created from a video track');

      this.done();
    }), onError);
}, 'verifies that an ImageCapture cannot be created out of an Audio Track');

</script>
